tutorials/rideindego/models/marshall/requirements.txt
pandas==1.2.3
click==7.1.2
loguru==0.5.3
python-dotenv==0.16.0
simplejson==3.17.0
requests==2.25.1
pyarrow==3.0.0